refactor: 도메인 모델을 src/api/*.types에서 src/models/로 분리
