+++ src/models/vote.ts
